Cross Validation from Scratch
Cross-validation is a technique used in machine learning to evaluate how well a model can generalize to new, unseen data.

Imagine you are trying to learn how to play a game, and you have a limited number of games to practice with. If you just practice with the same set of games over and over again, you may become really good at playing those specific games, but you won't necessarily be good at playing new games that you haven't seen before.

Similarly, when training a machine learning model, you don't want to just train it on the same set of data over and over again, because the model may become too specialized to that specific data and not perform well on new, unseen data.

Cross-validation addresses this issue by dividing the available data into multiple "folds" or subsets. The model is then trained on a subset of the data and evaluated on the remaining subset(s). This process is repeated multiple times, with different subsets of data used for training and evaluation each time.

By doing this, we can get a better estimate of how well the model will perform on new, unseen data, because we are essentially testing it on data that it hasn't seen before. This can help us identify any issues with overfitting or underfitting the data, and can also help us choose the best model parameters or architecture.

In this exercise, we will explore how to perform a cross validation without using any packages and what does it mean to a Machine Learning model